Update ent score every time a new comment is added.

+++ back-end/models/EnterpriseUser.py
@@ -24,7 +24,7 @@ class EnterpriseUser(db.Model):
     Contact = db.Column(db.String(256))
     Description = db.Column(db.String(1000))
 
-    CreditScore = db.Column(db.Integer)
+    CreditScore = db.Column(db.Float)
     
     aes = AES.new(extend_to_16(app.config['SECRET_KEY']), AES.MODE_ECB)
